+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Next version
 
+## Fixed
+
+- Add remote Swift packages to the Frameworks build phase https://github.com/tuist/XcodeProj/pull/487 by @kwridan
+- System library added to a group has empty path https://github.com/tuist/XcodeProj/pull/488 by @damirdavletov
+
 ## 7.1.0
 
 ### Added