+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 🚀 Check out the guidelines [here](https://github.com/xcodeswift/contributors/blob/master/CHANGELOG_GUIDELINES.md)
 
-## Next version
+## Next
 
 ## Fixed